+++ src/test/test-DigestAuthentication.ts
